#144 feat: 메일전송여부를 관리하는 매니저 객체 생성
