+++ src/main/java/codesquad/fineants/spring/api/portfolio_notification/PortfolioNotificationService.java
@@ -1,10 +1,8 @@
 package codesquad.fineants.spring.api.portfolio_notification;
 
-import java.time.Duration;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,6 +12,7 @@ import codesquad.fineants.domain.portfolio.PortfolioRepository;
 import codesquad.fineants.spring.api.errors.errorcode.PortfolioErrorCode;
 import codesquad.fineants.spring.api.errors.exception.NotFoundResourceException;
 import codesquad.fineants.spring.api.kis.manager.CurrentPriceManager;
+import codesquad.fineants.spring.api.portfolio_notification.manager.MailRedisManager;
 import codesquad.fineants.spring.api.portfolio_notification.request.PortfolioNotificationModifyRequest;
 import codesquad.fineants.spring.api.portfolio_notification.response.PortfolioNotificationModifyResponse;
 import lombok.RequiredArgsConstructor;
@@ -27,7 +26,7 @@ public class PortfolioNotificationService {
 
 	private final PortfolioRepository portfolioRepository;
 	private final MailService mailService;
-	private final RedisTemplate<String, String> redisTemplate;
+	private final MailRedisManager manager;
 	private final CurrentPriceManager currentPriceManager;
 
 	@Transactional
@@ -60,13 +59,13 @@ public class PortfolioNotificationService {
 	// 주기적으로 포트폴리오를 확인하여 목표수익금액에 도달하면 메일을 보낸다
 	@Scheduled(fixedRate = 120000L, initialDelay = 30000L)
 	public void notifyTargetGain() {
-		log.info("notifyTargetGain");
 		List<Portfolio> portfolios = portfolioRepository.findAll();
 		portfolios.forEach(portfolio -> portfolio.changeCurrentPriceFromHoldings(currentPriceManager));
 
 		List<Portfolio> reachedTargetGainPortfolios = portfolios.stream()
-			.filter(portfolio -> !hasMailSentHistoryForTargetGain(portfolio))
+			.filter(portfolio -> !manager.hasMailSentHistoryForTargetGain(portfolio))
 			.filter(Portfolio::getTargetGainIsActive)
+			.filter(portfolio -> !portfolio.isBudgetEmpty())
 			.filter(Portfolio::reachedTargetGain)
 			.collect(Collectors.toList());
 		log.info("포트폴리오 목표수익금액 도달 포트폴리오 : {}개", reachedTargetGainPortfolios.size());
@@ -76,27 +75,8 @@ public class PortfolioNotificationService {
 			String subject = String.format("%s 포트폴리오 목표수익금액 도달 안내 메일입니다", portfolio.getName());
 			String content = createMailContent(portfolio);
 			mailService.sendEmail(to, subject, content);
-			setMailSentHistoryForTargetGain(portfolio);
+			manager.setMailSentHistoryForTargetGain(portfolio);
 		});
-	}
-
-	private boolean hasMailSentHistoryForTargetGain(Portfolio portfolio) {
-		String value = redisTemplate.opsForValue().get(createMailSentHistoryKeyForTargetGain(portfolio));
-		return value != null;
-	}
-
-	private String createMailSentHistoryKeyForTargetGain(Portfolio portfolio) {
-		return String.format("targetGainMail:%d", portfolio.getId());
-	}
-
-	private boolean hasMailSentHistoryForMaximumLoss(Portfolio portfolio) {
-		String value = redisTemplate.opsForValue().get(createMailSentHistoryKeyForMaximumLoss(portfolio));
-		log.info("hasMailSentHistoryForMaximumLoss value={}", value);
-		return value != null;
-	}
-
-	private String createMailSentHistoryKeyForMaximumLoss(Portfolio portfolio) {
-		return String.format("MaximumLossMail:%d", portfolio.getId());
 	}
 
 	private String createMailContent(Portfolio portfolio) {
@@ -107,30 +87,16 @@ public class PortfolioNotificationService {
 		return String.join("\n", title, targetGain, maximumLoss, totalGain);
 	}
 
-	private void setMailSentHistoryForTargetGain(Portfolio portfolio) {
-		redisTemplate.opsForValue().set(
-			createMailSentHistoryKeyForTargetGain(portfolio),
-			Boolean.TRUE.toString(),
-			Duration.ofDays(1));
-	}
-
-	private void setMailSentHistoryMaximumLoss(Portfolio portfolio) {
-		redisTemplate.opsForValue().set(
-			createMailSentHistoryKeyForMaximumLoss(portfolio),
-			Boolean.TRUE.toString(),
-			Duration.ofDays(1));
-	}
-
 	// 1분마다 포트폴리오를 확인하여 최대손실금액에 도달하면 메일을 보낸다
 	@Scheduled(fixedRate = 120000L, initialDelay = 30000L)
 	public void notifyMaximumLoss() {
-		log.info("notifyMaximumLoss");
 		List<Portfolio> portfolios = portfolioRepository.findAll();
 		portfolios.forEach(portfolio -> portfolio.changeCurrentPriceFromHoldings(currentPriceManager));
 
 		List<Portfolio> reachedMaximumLossPortfolios = portfolios.stream()
-			.filter(portfolio -> !hasMailSentHistoryForMaximumLoss(portfolio))
+			.filter(portfolio -> !manager.hasMailSentHistoryForMaximumLoss(portfolio))
 			.filter(Portfolio::getMaximumIsActive)
+			.filter(portfolio -> !portfolio.isBudgetEmpty())
 			.filter(Portfolio::reachedMaximumLoss)
 			.collect(Collectors.toList());
 		log.info("포트폴리오 최대손실금액 도달 포트폴리오 : {}개", reachedMaximumLossPortfolios.size());
@@ -140,7 +106,7 @@ public class PortfolioNotificationService {
 			String subject = String.format("%s 포트폴리오 최대손실금액 도달 안내 메일입니다", portfolio.getName());
 			String content = createMailContent(portfolio);
 			mailService.sendEmail(to, subject, content);
-			setMailSentHistoryMaximumLoss(portfolio);
+			manager.setMailSentHistoryMaximumLoss(portfolio);
 		});
 	}
 }
